+++ src/export.ts
@@ -1,1 +1,1 @@
-export { socket as default, socket, port, status } from './layer/interface/api';
+export { storechannel, messagechannel } from './layer/interface/api';