+++ src/layer/data/store/event.ts
@@ -91,17 +91,17 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
     void this.events.save
       .monitor([], event => {
         switch (event.type) {
-          case EventStore.EventType.delete:
-          case EventStore.EventType.snapshot:
+          case EventStore.Event.Type.delete:
+          case EventStore.Event.Type.snapshot:
             void this.clean(event.key);
         }
       });
   }
   private readonly memory = new Observable<never[] | [K] | [K, string] | [K, string, string], void, UnsavedEventRecord<K, V> | SavedEventRecord<K, V>>();
   public readonly events = {
-    load: new Observable<never[] | [K] | [K, string] | [K, string, EventStore.EventType], EventStore.Event<K>, void>(),
-    save: new Observable<never[] | [K] | [K, string] | [K, string, EventStore.EventType], EventStore.Event<K>, void>(),
-    loss: new Observable<never[] | [K] | [K, string] | [K, string, EventStore.EventType], EventStore.Event<K>, void>(),
+    load: new Observable<never[] | [K] | [K, string] | [K, string, EventStore.Event.Type], EventStore.Event<K>, void>(),
+    save: new Observable<never[] | [K] | [K, string] | [K, string, EventStore.Event.Type], EventStore.Event<K>, void>(),
+    loss: new Observable<never[] | [K] | [K, string] | [K, string, EventStore.Event.Type], EventStore.Event<K>, void>(),
   };
   public readonly events_ = {
     update: new Observable<never[] | [K] | [K, string], UnsavedEventRecord<K, V> | SavedEventRecord<K, V>, void>(),
@@ -158,7 +158,7 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
             savedEvents
               // remove overridable event
               .reduceRight<SavedEventRecord<K, V>[]>((acc, e) =>
-                acc.length === 0 || acc[0].type === EventStore.EventType.put
+                acc.length === 0 || acc[0].type === EventStore.Event.Type.put
                   ? concat(acc, [e])
                   : acc
               , [])
@@ -191,7 +191,7 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
           const event: SavedEventRecord<K, V> = cursor.value;
           if (this.memory.refs([event.key, event.attr, sqid(event.id)]).length > 0) return void proc(null, err);
           void savedEvents.unshift(new SavedEventRecord(event.id, event.key, event.value, event.type, event.date));
-          if (event.type !== EventStore.EventType.put) return void proc(null, err);
+          if (event.type !== EventStore.Event.Type.put) return void proc(null, err);
           return void cursor.continue();
         }
       };
@@ -235,7 +235,7 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
     });
   }
   public has(key: K): boolean {
-    return compose(key, this.memory.reflect([key])).type !== EventStore.EventType.delete;
+    return compose(key, this.memory.reflect([key])).type !== EventStore.Event.Type.delete;
   }
   public get(key: K): V {
     if (!this.syncState.get(key)) {
@@ -247,22 +247,22 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
       .value;
   }
   public add(event: UnsavedEventRecord<K, V>, tx: IDBTransaction | void = this.tx): void {
-    assert(event.type === EventStore.EventType.snapshot ? tx : true);
+    assert(event.type === EventStore.Event.Type.snapshot ? tx : true);
     assert(!tx || tx.db.name === this.database && tx.mode === IDBTransactionMode.readwrite)
     void this.events_.access
       .emit([event.key, event.attr, event.type], new InternalEvent(event.type, IdNumber(0), event.key, event.attr));
-    if (!(event instanceof UnsavedEventRecord)) throw new Error(`LocalSocket: Cannot add a saved event: ${JSON.stringify(event)}`);
+    if (!(event instanceof UnsavedEventRecord)) throw new Error(`ClientChannel: Cannot add a saved event: ${JSON.stringify(event)}`);
     if (!this.syncState.get(event.key)) {
       void this.fetch(event.key);
     }
     switch (event.type) {
-      case EventStore.EventType.put: {
+      case EventStore.Event.Type.put: {
         void this.memory
           .off([event.key, event.attr, sqid(0)]);
         break;
       }
-      case EventStore.EventType.delete:
-      case EventStore.EventType.snapshot: {
+      case EventStore.Event.Type.delete:
+      case EventStore.Event.Type.snapshot: {
         void this.memory
           .refs([event.key])
           .filter(([[, , id]]) => id === sqid(0))
@@ -327,7 +327,7 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
         void this.events.loss.emit([event.key, event.attr, event.type], new EventStore.Event(event.type, IdNumber(0), event.key, event.attr, event.date)));
   }
   public delete(key: K): void {
-    return void this.add(new UnsavedEventRecord(key, <V>new EventStore.Value(), EventStore.EventType.delete));
+    return void this.add(new UnsavedEventRecord(key, <V>new EventStore.Value(), EventStore.Event.Type.delete));
   }
   private readonly snapshotCycle: number = 9;
   private snapshot(key: K): void {
@@ -351,7 +351,7 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
           const composedEvent = compose(key, savedEvents);
           if (composedEvent instanceof SavedEventRecord) return;
           switch (composedEvent.type) {
-            case EventStore.EventType.snapshot:
+            case EventStore.Event.Type.snapshot:
               // snapshot's date must not be later than unsaved event's date.
               return void this.add(
                 new UnsavedEventRecord(
@@ -360,10 +360,10 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
                   composedEvent.type,
                   savedEvents.reduce((date, e) => e.date > date ? e.date : date, 0)),
                 tx);
-            case EventStore.EventType.delete:
+            case EventStore.Event.Type.delete:
               return;
           }
-          throw new TypeError(`LocalSocket: Invalid event type: ${composedEvent.type}`);
+          throw new TypeError(`ClientChannel: Invalid event type: ${composedEvent.type}`);
         }
         else {
           return void cursor.continue();
@@ -389,11 +389,11 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
         else {
           const event: SavedEventRecord<K, V> = cursor.value;
           switch (event.type) {
-            case EventStore.EventType.put: {
+            case EventStore.Event.Type.put: {
               void cleanState.set(event.key, cleanState.get(event.key) || false);
               break;
             }
-            case EventStore.EventType.snapshot: {
+            case EventStore.Event.Type.snapshot: {
               if (!cleanState.get(event.key)) {
                 void cleanState.set(event.key, true);
                 void cursor.continue();
@@ -401,7 +401,7 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
               }
               break;
             }
-            case EventStore.EventType.delete: {
+            case EventStore.Event.Type.delete: {
               void cleanState.set(event.key, true);
               break;
             }
@@ -433,11 +433,9 @@ export abstract class EventStore<K extends string, V extends EventStore.Value> {
   }
 }
 export namespace EventStore {
-  export type EventType = Schema.EventType;
-  export const EventType = Schema.EventType;
   export class Event<K extends string> {
     constructor(
-      public readonly type: EventType,
+      public readonly type: Event.Type,
       public readonly id: IdNumber,
       public readonly key: K,
       public readonly attr: string,
@@ -445,6 +443,9 @@ export namespace EventStore {
     ) {
       void Object.freeze(this);
     }
+  }
+  export namespace Event {
+    export import Type = Schema.EventType;
   }
   export class Record<K extends string, V extends Value> extends UnsavedEventRecord<K, V> { }
   export class Value extends Schema.EventValue {
@@ -456,10 +457,10 @@ export {
 }
 
 const InternalEventType = {
-  ...EventStore.EventType,
+  ...EventStore.Event.Type,
   query: <'query'>'query'
 };
-type InternalEventType = EventStore.EventType | typeof InternalEventType.query;
+type InternalEventType = EventStore.Event.Type | typeof InternalEventType.query;
 class InternalEvent<K extends string> {
   constructor(
     public readonly type: InternalEventType,
@@ -487,7 +488,7 @@ export function compose<K extends string, V extends EventStore.Value>(
   return group(events)
     .map(events =>
       events
-        .reduceRight(compose, new UnsavedEventRecord(key, <V>new EventStore.Value(), EventStore.EventType.delete, 0)))
+        .reduceRight(compose, new UnsavedEventRecord(key, <V>new EventStore.Value(), EventStore.Event.Type.delete, 0)))
     .reduce(e => e);
 
   function group(events: E[]): E[][] {
@@ -508,12 +509,12 @@ export function compose<K extends string, V extends EventStore.Value>(
     assert(target instanceof UnsavedEventRecord || target instanceof SavedEventRecord);
     assert(source instanceof UnsavedEventRecord || source instanceof SavedEventRecord);
     switch (source.type) {
-      case EventStore.EventType.put:
+      case EventStore.Event.Type.put:
         return source.value[source.attr] !== void 0
           ? new UnsavedEventRecord(
             source.key,
             Object.assign(new EventStore.Value(), target.value, source.value),
-            EventStore.EventType.snapshot)
+            EventStore.Event.Type.snapshot)
           : new UnsavedEventRecord(
             source.key,
             Object.keys(target.value)
@@ -522,12 +523,12 @@ export function compose<K extends string, V extends EventStore.Value>(
                 value[prop] = target[prop];
                 return value;
               }, <V>new EventStore.Value())
-            , EventStore.EventType.snapshot);
-      case EventStore.EventType.snapshot:
+            , EventStore.Event.Type.snapshot);
+      case EventStore.Event.Type.snapshot:
         return source;
-      case EventStore.EventType.delete:
+      case EventStore.Event.Type.delete:
         return source;
     }
-    throw new TypeError(`LocalSocket: Invalid event type: ${source}`);
+    throw new TypeError(`ClientChannel: Invalid event type: ${source}`);
   }
 }