+++ src/layer/domain/indexeddb/api.ts
@@ -1,2 +1,2 @@
-export { Socket } from './service/socket';
+export { Channel as StoreChannel } from './service/channel';
 export { event, IDBEventType } from './service/event';