+++ src/layer/domain/indexeddb/model/channel/access.ts
@@ -48,7 +48,7 @@ export class AccessStore<K extends string> extends KeyValueStore<K, AccessRecord
 
     void access
       .monitor([], ({key, type}) =>
-        type === EventStore.EventType.delete
+        type === EventStore.Event.Type.delete
           ? void this.delete(key)
           : void this.set(key, new AccessRecord(key, Date.now()))
       );