+++ src/layer/domain/webstorage/api.ts
@@ -1,3 +1,3 @@
 export { localStorage, sessionStorage, supportWebStorage } from '../../infrastructure/webstorage/api';
 export { events } from './service/event';
-export { Port, PortEvent as WebStorageEvent, PortEventType as WebStorageEventType } from './service/port';
+export { Channel as MessageChannel, ChannelObject as MessageChannelObject, ChannelEvent as MessageChannelEvent } from './service/channel';