+++ src/layer/domain/webstorage/service/channel.test.ts
@@ -1,9 +1,9 @@
-import { LocalPortObject } from '../../../../../';
-import { Port } from './port';
+import { MessageChannelObject } from '../../../../../';
+import { Channel } from './channel';
 
-describe('Unit: layers/domain/webstorage/service/port', () => {
+describe('Unit: layers/domain/webstorage/service/channel', () => {
   describe('spec', () => {
-    interface DAO extends LocalPortObject {
+    interface DAO extends MessageChannelObject {
     }
     class DAO {
       n = 0;
@@ -21,13 +21,13 @@ describe('Unit: layers/domain/webstorage/service/port', () => {
     });
 
     it('singleton', () => {
-      assert(new Port('test', sessionStorage, factory) === new Port('test', sessionStorage, factory));
-      new Port('test', sessionStorage, factory).destroy();
+      assert(new Channel('test', sessionStorage, factory) === new Channel('test', sessionStorage, factory));
+      new Channel('test', sessionStorage, factory).destroy();
     });
 
     it('make/destroy', () => {
       assert(sessionStorage.getItem('test') === null);
-      const repo = new Port('test', sessionStorage, factory);
+      const repo = new Channel('test', sessionStorage, factory);
       const dao = repo.link();
       assert(dao.__key === 'test');
       assert(dao.n === 0);
@@ -45,15 +45,15 @@ describe('Unit: layers/domain/webstorage/service/port', () => {
     it('remake', () => {
       assert(sessionStorage.getItem('test') === null);
       assert.equal(
-        new Port('test', sessionStorage, factory).link(),
-        new Port('test', sessionStorage, factory).link()
+        new Channel('test', sessionStorage, factory).link(),
+        new Channel('test', sessionStorage, factory).link()
       );
-      new Port('test', sessionStorage, factory).destroy();
+      new Channel('test', sessionStorage, factory).destroy();
       assert(sessionStorage.getItem('test') === null);
     });
 
     it('update', () => {
-      const repo = new Port('test', sessionStorage, factory);
+      const repo = new Channel('test', sessionStorage, factory);
       const dao = repo.link();
       assert(dao.n === 0);
       dao.n = 1;